feat: add useSortableHeader composable
