+++ packages/vlossom/src/components/vs-table/composables/useTableSort.ts
@@ -1,34 +1,8 @@
 import { Ref } from 'vue';
-import { TableHeader, TableItem, SortType } from '../types';
+import { TableItem, SortType } from '../types';
 
-export function useTableSort(sortTypes: Ref<{ [key: string]: SortType }>) {
-    function updateSortType({ key, sortable }: TableHeader) {
-        if (!sortable) {
-            return;
-        }
-        Object.keys(sortTypes.value).forEach((sortKey: string) => {
-            if (key === sortKey) {
-                const type = sortTypes.value[sortKey];
-                sortTypes.value[sortKey] = (type + 1) % 3;
-            } else {
-                sortTypes.value[sortKey] = SortType.NONE;
-            }
-        });
-    }
-
-    function getSortIcon(type: SortType) {
-        switch (type) {
-            case SortType.ASCEND:
-                return 'expandMore';
-            case SortType.DESCEND:
-                return 'expandLess';
-            case SortType.NONE:
-            default:
-                return 'unfoldMore';
-        }
-    }
-
-    function getSortedItems(items: TableItem[]) {
+export function useTableSort() {
+    function getSortedItems(items: TableItem[], sortTypes: Ref<{ [key: string]: SortType }>) {
         const sortKey = Object.keys(sortTypes.value).find((key) => {
             return sortTypes.value[key] !== SortType.NONE;
         });
@@ -47,8 +21,6 @@ export function useTableSort(sortTypes: Ref<{ [key: string]: SortType }>) {
     }
 
     return {
-        updateSortType,
-        getSortIcon,
         getSortedItems,
     };
 }
